Bug 573004 persist generated key pairs
refactor prior invasion
